requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
iexfinance==0.4.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
urllib3==1.24.1
whitenoise==4.1.2
